--- mito_utils/__init__.py
@@ -1,5 +1,0 @@
-from . import (
-                classification, clustering, colors, diagnostic_plots,
-                dimred, distances, heatmaps_plots, kNN, plotting_base, 
-                preprocessing, utils, phylo, phylo_plots
-)
